+++ src/Lykke.Service.Chainalysis.Mock.AzureRepositories/Dto/UserRepoData.cs
@@ -8,9 +8,9 @@ namespace Lykke.Service.ChainalysisMock.AzureRepositories.Dto
     public class UserRepoData : IUserRepoData
     {
         public string UserId { get; set; }
-        public int LastActivity { get; set; }
+        public long LastActivity { get; set; }
         public string Score { get; set; }
-        public int ScoreUpdatedDate { get; set; }
-        public int CreatedDate { get; set; }
+        public long ScoreUpdatedDate { get; set; }
+        public long CreatedDate { get; set; }
     }
 }
