+++ src/Lykke.Service.Chainalysis.Mock.Contracts/UserDataModel.cs
@@ -3,8 +3,8 @@
     public class UserDataModel
     {
         public string UserId { get; set; }
-        public int LastActivity { get; set; }
+        public long LastActivity { get; set; }
         public string Score { get; set; }
-        public int ScoreUpdatedDate { get; set; }
+        public long ScoreUpdatedDate { get; set; }
     }
 }
