+++ src/Lykke.Service.Chainalysis.Mock.Services/Dto/UserDetails.cs
@@ -12,11 +12,11 @@ namespace Lykke.Service.ChainalysisMock.Services.Dto
             ExposureDetails = new List<IUserExplosureDetails> {new UserExplosureDetails()};
         }
         public string UserId { get; set; }
-        public int CreationDate { get; set; }
+        public long CreationDate { get; set; }
         public string Comment { get; set; }
-        public int LastActivity { get; set; }
+        public long LastActivity { get; set; }
         public string Score { get; set; }
-        public int ScoreUpdatedDate { get; set; }
+        public long ScoreUpdatedDate { get; set; }
         public List<IUserExplosureDetails> ExposureDetails { get; set; }
     }
 }
