+++ src/Lykke.Service.Chainalysis.Mock.Services/Dto/UserExplosureDetails.cs
@@ -17,9 +17,9 @@ namespace Lykke.Service.ChainalysisMock.Services.Dto
 
 
         public string Category { get; set; }
-        public int SentIndirectExposure { get; set; }
-        public int SentDirectExposure { get; set; }
-        public int ReceivedIndirectExposure { get; set; }
-        public int ReceivedDirectExposure { get; set; }
+        public long SentIndirectExposure { get; set; }
+        public long SentDirectExposure { get; set; }
+        public long ReceivedIndirectExposure { get; set; }
+        public long ReceivedDirectExposure { get; set; }
     }
 }
